+++ packages/components/navbar/src/js/menu/navbar-menu.component.js
@@ -14,6 +14,7 @@ export default {
     menuLinks: '<links',
     menuName: '@name',
     placement: '@?',
+    variant: '@?',
     fixed: '<?',
   },
   transclude: {